+++ frontend/src/components/RecoverVaultDialog.vue
@@ -60,9 +60,11 @@ import { VaultKeys } from '../common/crypto';
 import userdata from '../common/userdata';
 
 class FormValidationFailedError extends Error {
+
   constructor() {
     super('The form is invalid.');
   }
+
 }
 
 const { t } = useI18n({ useScope: 'global' });